Move tailscaled to cli and expose Run, similiar to cmd/tailscale
