+++ cmd/tailscaled/cli/manifest_windows_amd64.syso
