+++ cmd/tailscaled/cli/tailscaled_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
-package main // import "tailscale.com/cmd/tailscaled"
+package cli // import "tailscale.com/cmd/tailscaled"
 
 import (
 	"testing"
