+++ cmd/tailscaled/cli/windows-manifest.xml
