+++ cmd/tailscaled/cli/with_cli.go
@@ -3,7 +3,7 @@
 
 //go:build ts_include_cli
 
-package main
+package cli
 
 import (
 	"fmt"
